+++ uses.go
@@ -145,7 +145,7 @@ func ExecuteUses(ctx context.Context, store *storage.Store, uses string, with Wi
 		if err != nil {
 			return err
 		}
-		
+
 		exists, err := store.Exists(desc)
 		if err != nil {
 			return err
